Refactor: `Member`와 `Profile` 컨트롤러와 서비스 분리
- `Member`와 `Profile` 컨트롤러와 서비스 분리
- 관련된 명칭 변경
