+++ src/main/java/com/pinterest/service/MemberService.java
@@ -1,10 +1,8 @@
 package com.pinterest.service;
 
 import com.pinterest.config.CustomUserDetails;
-import com.pinterest.domain.FileEntity;
 import com.pinterest.domain.Member;
 import com.pinterest.domain.MemberRole;
-import com.pinterest.dto.MemberDto;
 import com.pinterest.dto.ProfileDto;
 import com.pinterest.dto.request.JoinRequest;
 import com.pinterest.repository.MemberRepository;
@@ -15,7 +13,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
@@ -26,7 +23,6 @@ import java.util.List;
 @Slf4j
 public class MemberService implements UserDetailsService {
 
-    private final FileService fileService;
     private final MemberRepository memberRepository;
     private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
@@ -60,23 +56,5 @@ public class MemberService implements UserDetailsService {
         return memberRepository.findById(memberId)
                 .map(ProfileDto::from)
                 .orElseThrow(() -> new EntityNotFoundException("프로필이 없습니다."));
-    }
-
-    @Transactional
-    public void updateMember(Long profileId, MemberDto dto, MultipartFile file) {
-        try {
-            Member member = memberRepository.getReferenceById(profileId);
-
-            if (member.getEmail().equals(dto.getEmail())) {
-                if (file.isEmpty()) {
-                    member.update(dto.getNickname());
-                } else {
-                    FileEntity fileEntity = fileService.saveFile(file);
-                    member.update(dto.getNickname(), fileEntity);
-                }
-            }
-        } catch (EntityNotFoundException e) {
-            log.warn("프로필 업데이트 실패. 프로필을 찾을 수 없습니다.");
-        }
     }
 }
