+++ src/main/resources/templates/base/header.html
@@ -37,7 +37,7 @@
         <a th:href="@{/members/signup}" class="rounded-full bg-gray-400 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-gray-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-gray-600">가입하기</a>
       </th:block>
       <th:block sec:authorize="isAuthenticated()">
-        <a th:href="@{/members/{email}(email=${#authentication.name})}" class="rounded-full bg-red-600 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-red-800 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-red-600">마이페이지</a>
+        <a th:href="@{/profiles/{email}(email=${#authentication.name})}" class="rounded-full bg-red-600 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-red-800 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-red-600">마이페이지</a>
         <a th:href="@{/logout}" class="rounded-full bg-gray-400 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-gray-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-gray-600">로그아웃</a>
       </th:block>
     </div>
